refactor(workbench): move `stringifyError` function to a dedicated file
