+++ projects/scion/workbench/src/lib/common/stringify-error.util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2022 Swiss Federal Railways
+ * Copyright (c) 2018-2023 Swiss Federal Railways
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -11,9 +11,9 @@
 /**
  * Returns the error message if given an error object, or the `toString` representation otherwise.
  */
-export function stringifyError(error: any): string {
+export function stringifyError(error: unknown): string {
   if (error instanceof Error) {
     return error.message;
   }
-  return error?.toString();
+  return `${error}`;
 }
